requirements.txt
django==3.0.5
djangorestframework==3.12.4
djoser=2.1.0
pillow
django-filter
flake8